+++ clang-tools-extra/unittests/clang-tidy/ObjCModuleTest.cpp
@@ -1,0 +1,21 @@
+//===---- ObjCModuleTest.cpp - clang-tidy ---------------------------------===//
+//
+//                     The LLVM Compiler Infrastructure
+//
+// This file is distributed under the University of Illinois Open Source
+// License. See LICENSE.TXT for details.
+//
+//===----------------------------------------------------------------------===//
+
+#include "ClangTidyTest.h"
+#include "gtest/gtest.h"
+
+namespace clang {
+namespace tidy {
+namespace test {
+
+// TODO(D39142) Add unit tests for the ObjC module here once a check lands.
+
+} // namespace test
+} // namespace tidy
+} // namespace clang